Update README.md
About the data
The dataset is used for this competition is synthetic (and generated using a CTGAN), but based on a real dataset. The original dataset deals with predicting the amount of an insurance claim. Although the features are anonymized, they have properties relating to real-world features.

How to spend the next two weeks: keep learning!
Aim to make at least one submission per day. Even if you are low on time, you can learn a lot from this exercise, and it is very motivating to gradually move up the leaderboard!

As described in the Getting Started Guide above, the best way to keep learning from the community is through the Discussion tab and Code tab above.

Through the Discussion tab, you can chat with users about ideas or doubts that you have, or find a team to work with.
Through the Code tab, you can share your code and also learn from code that other Kagglers have shared with the community.
If you're interested in learning new skills through courses, check out the Intro to Deep Learning and Feature Engineering courses, which teach new techniques that could be relevant for working with the competition data.

The data for this competition is closely related to the Tabular Playground Series - Feb 2021 competition. Refer to the Code tab and the Discussion tab for ideas of algorithms that you might like to try, that may also perform well in this competition!

Important Note
You are only eligible to join this competition if you are registered for the 30 Days of ML challenge and have received an invite link to your personal email.

Please do not share the invitation link that you have received by email with anyone else.
In order to be eligible to win a prize, you must be registered for the 30 Days of ML challenge.